Refactored nut_gui.py * whitelisted PyQt5 for pylint * added nut_gui.py for pylinting
