+++ gui/table_model.py
@@ -8,7 +8,9 @@ from nut import Print
 
 
 def _format_size(num, suffix='B'):
-	for unit in ['','Ki','Mi','Gi','Ti','Pi','Ei','Zi']:
+	if num is None:
+		return ''
+	for unit in ['', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi']:
 		if abs(num) < 1024.0:
 			return "%3.1f %s%s" % (num, unit, suffix)
 		num /= 1024.0
