+++ tasks.py
@@ -23,7 +23,7 @@ def test(c):
 
 @task
 def lint(c):
-	run_arg = "pylint -j 4 nut/Config.py tests/ nut/Nsps.py nut/Hex.py"
+	run_arg = "pylint -j 4 nut/Config.py tests/ nut/Nsps.py nut/Hex.py nut_gui.py"
 	if os.name == 'nt': # Windows
 		c.run(run_arg)
 	else:
